[integ-tests-framework] Move push_result_to_dynamodb to tests/common/utils
The SSM automation test pipeline deletes tests/performance_tests before running tests, breaking nccl_common.py's top-level import from that package.

Moreover, The function is a generic DynamoDB reporting utility with no performance-test-specific logic, so tests/common/utils is a more appropriate home regardless.
